+++ fuzz/corpus/kudu_sql/seed3.sql
@@ -1,0 +1,1 @@
+SELECT * FROM t WHERE name = 'O''Brien' -- comment